requirements.txt
astroid==3.2.2
autopep8==2.3.0
beautifulsoup4==4.12.3
Brotli==1.0.9
certifi==2024.6.2
charset-normalizer==3.3.2
dill==0.3.8
exceptiongroup==1.2.1
filelock==3.13.1
fsspec==2024.6.1
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
huggingface-hub==0.23.4
idna==3.7
iniconfig==2.0.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.3
mccabe==0.7.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.1
parso==0.8.4
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.12.0
pylint==3.2.3
PySocks==1.7.1
pytest==8.2.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
setuptools==69.5.1
smmap==5.0.1
soupsieve==2.5
sympy==1.12
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.5
torch==2.0.1
tqdm==4.66.4
tree-sitter-languages==1.10.2
torch==2.0.1
numpy==1.23.3
transformers==4.26.1
triton==2.3.1
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
